utils: Move struct enum_def and enum_val to auto-args.h
This creates auto-args.h and move struct enum_def and enum_val to the
header file for external references.
